fix : transaction domain(demical -> double), convertGasToEther)

+++ src/main/java/e1/i3/e1i3/domain/transaction/Transaction.java
@@ -34,6 +34,5 @@ public class Transaction extends BaseTime {
     private Long timeStamp;
 
     @Column
-    private BigDecimal gasUsed;
-
+    private double gasUsed;
 }
